+++ Source/Parsec.swift
@@ -253,20 +253,6 @@ public struct APIRelationship {
     }
 }
 
-public enum LogLevel: Int {
-    case error
-    case warning
-    case info
-    case debug
-    case verbose
-}
-
-public protocol Logger {
-    func log(_ level: LogLevel, _ text: @autoclosure () -> String)
-    var logUnknown: Bool {get}
-    var logMissing: Bool {get}
-}
-
 /// *Parsec*
 public class Parsec {
